+++ prettier/plugin/src/getters/printers.getter.ts
@@ -1,5 +1,5 @@
 import type { GetPrintersReturn } from '../interfaces/index.js'
 
-import { printers }               from '../printers.js'
+import { printers }               from '../printers/index.js'
 
 export const getPrinters = async (): GetPrintersReturn => printers
