+++ prettier/plugin/src/parsers.ts
@@ -1,6 +1,5 @@
 /* eslint-disable @typescript-eslint/no-unsafe-call */
 
-import type { ImportDeclaration }             from '@babel/types'
 import type { Parser }                        from 'prettier'
 
 import * as babel                             from 'prettier/plugins/babel'
@@ -12,35 +11,6 @@ import { preprocess as importSortPreprocess } from './import-sort/index.js'
 const preprocess: Parser['preprocess'] = (source, options): string =>
   importSortPreprocess(source, options)
 
-type ImportSpecifier = ImportDeclaration['specifiers'][number]
-
-const isUpperCase = (value: string | undefined): boolean => {
-  const firstCharacter = value?.[0]
-
-  return Boolean(firstCharacter && firstCharacter === firstCharacter.toUpperCase())
-}
-
-const getImportSpecifierName = (specifier: ImportSpecifier): string | undefined =>
-  specifier.local.name
-
-const getImportSpecifierRank = (specifier: ImportSpecifier): number => {
-  if (specifier.type === 'ImportNamespaceSpecifier') {
-    return 0
-  }
-
-  if (specifier.type === 'ImportSpecifier') {
-    return isUpperCase(getImportSpecifierName(specifier)) ? 1 : 2
-  }
-
-  return isUpperCase(getImportSpecifierName(specifier)) ? 3 : 4
-}
-
-const sortSplitImportSpecifiers = (specifiers: Array<ImportSpecifier>): Array<ImportSpecifier> =>
-  specifiers
-    .map((specifier, index) => ({ index, rank: getImportSpecifierRank(specifier), specifier }))
-    .sort((left, right) => left.rank - right.rank || left.index - right.index)
-    .map(({ specifier }) => specifier)
-
 const parse: Parser['parse'] = async (source, { plugins }) => {
   // @ts-expect-error parser options type is wider at runtime than @types/prettier declares
   const program = typescript.parsers.typescript.parse(source, { plugins })
@@ -51,19 +21,16 @@ const parse: Parser['parse'] = async (source, { plugins }) => {
 
   nodes.forEach((node, nodeIndex: number) => {
     if (node.type === 'ImportDeclaration') {
-      const importNode = node as ImportDeclaration
-
-      if (importNode.specifiers.length > 1) {
+      if (node.specifiers.length > 1) {
         const index = bodyLength - nodeIndex - 1
 
         program.body.splice(index, 1)
 
-        const splitSpecifiers = sortSplitImportSpecifiers(importNode.specifiers)
-
-        splitSpecifiers.forEach((specifier, specifierIndex: number) => {
+        node.specifiers.forEach((_: unknown, specifierIndex: number) => {
           program.body.splice(index + specifierIndex, 0, {
-            ...importNode,
-            specifiers: [specifier],
+            ...node,
+            // eslint-disable-next-line @typescript-eslint/no-shadow
+            specifiers: node.specifiers.filter((_: unknown, i: number) => specifierIndex === i),
           })
         })
       }
